+++ module-domain-menu/src/main/java/domain/menu/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package domain.menu;
 
 
 import common.entity.BaseIdEntity;
