+++ module-domain-menu/src/main/java/domain/menu/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package domain.menu;
 
 import common.entity.Price;
 import org.hibernate.annotations.BatchSize;
