+++ module-domain-menu/src/main/java/domain/menu/MenuValidationException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.application;
+package domain.menu;
 
 
 import common.application.ValidationException;
