+++ settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'kitchenpos'
 include 'common'
 include 'module-api-config'
+include 'module-domain-menu'
+
