+++ src/main/java/kitchenpos/Application.java
@@ -5,12 +5,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 import javax.sql.DataSource;
 
-@EnableJpaRepositories
+@EnableJpaRepositories(basePackages = {"domain.menu", "kitchenpos.order"})
+@EntityScan(basePackages = {"domain.menu", "kitchenpos.order"})
 @EnableJpaAuditing
 @SpringBootApplication
 public class Application implements CommandLineRunner {
